Build the Communicator interface to allow GridFragment to pass information back to the activity that spawned it. For now, it just accepts an integer to pass the position back to the ActivityMain.
GridFragment - Implemented the Communicator interface. Built an interface reference variable, and pointed it at the calling activity. Modified the onItemClick to call communicator.respond(position), to send the position data back to the spawning activity.

MainActivity - implements the Communicator interface, and therefore must use the respond() method. Respond now logs the item clicked in the fragment. Data is being passed from the fragment to the activity.