MainActivity - Override onCreateOptionsMenu to inflate menu_main menu resource. Override onOptionsItemSelected to capture selections and do something about them. Right now, there's only a settings option in the menu, which launches a new MenuActivity via intent.
MenuActivity - blows up menu_activity layout, which is nothing but a grey background at the moment.

menu_activity.xml - the layout that will eventually become options you're presented with when you pick Settings from the Action Bar

menu_main - the layout that defines the choices in the Action Bar. So far, just Settings.