DatabaseAdapter handles all database operations. Methods to allow outside apps to open and close the database, do a bulk insert or insert a single item.
DBManager should be deleted. All of this functionality has been migrated to an inner class of DBManager.

GridFragment isn't doing much at the moment. Most of the meaningful code has migrated off of the fragment and into other classes. Execution has been commented out in advance of a peeling back of this deprecated code.

MainActivity - onCreate calls updateMovies(), which gets migrated from the GridFragment. Essentially just kicks off the download. The AsyncTask has been reworked a bit. URL download is the same, and rawData gets sent to reformatMovieData(). All of the failures now return false, A successful download and parsing will return true to onPostExecute. reformatMovieData() has been reworked to send each movieobject to the itemInsert method on the DatabaseAdapter. This is lagging a bit. I will probably revert back to dropping everything into an ArrayList<MovieObject> and sending that to the bulkAdd method.