DatabaseAdapter gets a new method - getDBCount(), which does exactly what the name implies. It's just there to prove I can stand up the database and read something from the GridFragment, which will have to use the database to populate the GridView.
DBManager no longer serves a purpose and has been removed.

GridFragment gets a reference to the database, a couple of methods to open and close it, and does a quick open/query/close at onResume(); The query is working and dumping the expected values in to the log.

MainActivity pulled some simple logging now that the functionality is all worked out.

This is a pretty solid milestone. The web/parse/drop into database is working as part of the main activity. The fragment can now access that data. From here, I'll need to disallow duplicate entries in the database. (UNIQUE sql magic or query and check before insert?) I will also need to design a custom item_view for the gridview itself. Something that can reflect whether the movies has been marked as a favorite. Finally, I'll need to attach a cursor adapter to the gridview. It'll need to take a look at the popular/rated preference to figure out which views to display, and also look at the favorites table to see if it needs to mark the poster.