Altered TMDBContract to replace duplicate entries, keyed on the movieID. This will keep the database from needlessly expanding.
MainActivity - Moved a few things around. There were some autogenerated Override methods that came as part of the last AS upgrade. Just moved those to the bottom for now. Moved the API key to the top of the MainActivity just to make life a little easier.

Commented the logging out of the Database adapter and a number of other locations. Things are working. No need to spam the logs.