Quick refactor to get rid of a few dead ends and move some things off of the MainActivity.
Since we've moved to the ContentProvider model for data access, the DatabaseAdapter class is no longer necessary.

GridFragment - removed all calls to the database adapter and anything that accesses the database from the fragment.

InternetDownloadTask. I moved the updatemovies() method and the FetchMoviesTask class to a new class InternetDownloadTask.

MainActivity - download and database population logic gets moved off of the main activity and into its own class. MainActivity simply instatiates InternetDownloadTask onCreate(), and then calls the updatemovies() method on that object.