Display Activity - technically eliminated, but actually just renamed and refactored to DisplayFragment.
DisplayFragment - Now receives a long from Communicator and pulls the information from the database using that number as the _ID field in the search. Still some work to do on the fragment wrt lifecycle.

Communicator - slight rework of the interface to force respond to accept a long as opposed to an arraylist.