Communicator - reduced to nothing more than a callback. Doesn't pass data anymore.
DisplayFragment - pulled the deprecated passedMovie stuff

GridFragment - Clicklistener now writes the user's choice to the db

gvCursorAdapter - tried introducing a placeholder image to see if it would have an impact on the flickering in the gridview. No change. This is a bug. I'm missing something!

MainActivity - built some logic to maintain the right fragment in front of the user as activities are destroyed and recreated on rotation.