DisplayExtras - Obect to keep track of reviews and trailers.
DisplayFragment - Now looks at the requested movie in the database, and seeks out information on Trailers and Reviews. These are unfortunately two web queries, so there doesn't seem the be any way to implement this other than to kick it off twice. Extras are downloaded, parsed, and dropped in an arraylist of DisplayExtras. Just keeping track of what type it is (Review or Trailer) and the associated URL.

All of the information is now coming down and getting dropped into an arraylist. Need to build the UX element and adapter.