DisplayFragment - Implemented a default case for moviequery as, with a two pane display, it's possible to have a DisplayFragment instantiated before a user has made a selection. This lead to an invalid request, which would crash the application.
MainActivity - Eliminated all of the logic that lead to backstacking multiple fragments. Added logic to detect whether the app was running on a screen large enough to handle 2 panes and, if so, populate the second pane with the display fragment.

activity_main (both standard and sw-600 now exist) - In both cases, activity_main hosts a static fragment that points to GridFragment. sw-600 contains a FrameLayout with id display_activity to provide a place for MainActivity to fire up a DisplayFragment. The standard activity_main does not include this fragment. Rather, display_activity contains a FrameLayout id display_activity, which DisplayActivity, not MainActivity, will populate in one pane mode.