Communicator gets a new method, likeButton();
DisplayActivity must implement communicator, because DisplayFragment uses communicator, and gets a reference back to the calling activity when it does so. DisplayActivity's respond() and likeButton() methods are empty, as in single pane mode no UI change needs to be made to the GridFragment.

DisplayFragment uses the Communicator interface to get a reference back to the calling activity. It uses this callback when the Like button is clicked. What that button does depends upon the calling activity's implementation of the method. DisplayActivity, which only operates in single pane mode, does nothing when this method is called. MainActivity, if it sees the device is in two pane mode, will reload the GridFragment, refreshing the display. The refreshed view will reflect the most up to date "liked" status.

MainActivity - respond() method updated to handle 1 or 2 pane displays. 1 pane, respond launches DisplayActivity by intent. 2 pane, respond sticks a DisplayFragment into id:display_activity. likeButton() resets the gridFragment.